fix: Round-3 — match galois_upper spot check to the proptest form
Drop the `as i32` cast in `int_uint_galois_at_mid_positive`. The
right-hand side compares an `i8` to an `i8`; widening to `i32` was
unnecessary and made the spot check express a different shape than
the `single_sided_props!` proptest, which tests the law on the
native types directly.

+++ src/conn/uint.rs
@@ -148,7 +148,7 @@ mod tests {
         for (a, b) in [(50i8, 100u16), (50i8, 49u16), (100i8, 100u16)] {
             assert_eq!(
                 I08U16.ceil(a) <= b,
-                a as i32 <= I08U16.inner(b) as i32,
+                a <= I08U16.inner(b),
                 "I08U16 galois_upper @ ({a}, {b})"
             );
         }
